+++ Adj/Program/Datastructure/Implementation/Tree.hs
@@ -1,7 +1,11 @@
 module Adj.Program.Datastructure.Implementation.Tree where
 
-import Adj.Auxiliary (type (=!??=))
+import Adj.Algebra.Set (type (=:*:=))
 import Adj.Program.Primitive.Maybe (Maybe)
 import Adj.Program.Primitive.Generation (Construction)
 
 type Tree = Construction
+
+type Binary tree = Tree (Maybe =:*:= Maybe)
+
+type Ternary tree = Tree (Maybe =:*:= Maybe =:*:= Maybe)